refactor: move services to db folder
