+++ src/server/db/services/CommentCrud.ts
@@ -1,5 +1,5 @@
 import { BaseCrud } from './BaseCrud';
-import { CommentAttributes, CommentCreationAttributes, Comment } from '../db/models/Comment';
+import { CommentAttributes, CommentCreationAttributes, Comment } from '../models/Comment';
 
 class CommentCrud extends BaseCrud<CommentAttributes, CommentCreationAttributes> {
   constructor() {
