--- src/server/orm/models/TestCrud/TestCrud.ts
@@ -1,13 +1,0 @@
-import { Model } from 'sequelize';
-import { TestCrudTableAttributes, TestCrudTableCreationAttributes } from './types';
-
-export class TestCrudModel
-  extends Model<TestCrudTableAttributes, TestCrudTableCreationAttributes>
-  implements TestCrudTableAttributes
-{
-  id!: number;
-
-  text!: string;
-
-  description!: string | null;
-}
