--- src/server/orm/models/TestCrud/index.ts
@@ -1,2 +1,0 @@
-export { TestCrudModel } from './TestCrud';
-export { TestCrudTableAttributes, TestCrudTableCreationAttributes } from './types';
