--- src/server/orm/models/TestCrud/types.ts
@@ -1,9 +1,0 @@
-import { Optional } from 'sequelize';
-
-export interface TestCrudTableAttributes {
-  id: number;
-  text: string;
-  description: string | null;
-}
-
-export interface TestCrudTableCreationAttributes extends Optional<TestCrudTableAttributes, 'id'> {}
